# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: NOURA API
  version: 0.1.0
  description: |
    NOURA (working name) mobile API. This document is the contract authority for the API and the
    generated Dart client (blueprint §3, §11).

    Conventions:
    - Base path `/v1`; JSON with snake_case fields; UUID identifiers; ISO-8601 timestamps; metric units.
    - Bearer authentication with a Supabase access token, except health endpoints and the
      provider webhook (which uses its own authorization header).
    - Success envelope: `{"data": {...}, "meta": {"request_id": "..."}}`.
    - Error envelope: `{"error": {"code", "message", "field_errors", "retryable"}, "meta": {...}}`.
    - Missing and non-owned resources both return 404.
    - Mutations require an `Idempotency-Key` header; edits require `expected_revision` and return
      409 `REVISION_CONFLICT` when stale.
    - Mutation request bodies reject unrecognized fields (`additionalProperties: false`).

    Every operation carries `x-noura-milestone` (the milestone that implements it) and
    `x-noura-status` (`implemented` or `planned`). Planned operations are contract drafts only: the
    API does not register them, and they may be refined (additively where possible) by their milestone.
  license:
    name: Proprietary
    url: https://example.invalid/license
servers:
  - url: http://localhost:8080
    description: Local development API
security:
  - bearerAuth: []
...
paths:
...
  /v1/media/upload-slots:
    post:
      operationId: createUploadSlot
      summary: Reserve an owner-scoped signed upload slot.
      tags: [media]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadSlotRequest' }
      responses:
        '201':
          description: Owner-scoped signed upload slot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSlotResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/QuotaExceeded' }
  /v1/media/{id}/complete:
    post:
      operationId: completeUpload
      summary: Confirm an upload and verify the stored bytes.
      tags: [media]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Verified media status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaAssetResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/media/{id}/download:
    get:
      operationId: getMediaDownload
      summary: Short-lived signed download link for owned media.
      tags: [media]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Short-lived signed download link.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaDownloadResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/media/{id}:
    delete:
      operationId: deleteMedia
      summary: Delete owned media.
      tags: [media]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Deletion status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaAssetResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/meal-scans:
    post:
      operationId: createMealScan
      summary: Start recognition for an uploaded meal image.
      tags: [meals]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMealScanRequest' }
      responses:
        '202':
          description: Scan job accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MealScanAcceptedResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/QuotaExceeded' }
  /v1/meal-scans/{id}:
    get:
      operationId: getMealScan
      summary: Meal scan state and recognition candidates.
      tags: [meals]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Scan state and recognition candidates.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MealScanResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/meal-scans/{id}/confirmed-items:
    put:
      operationId: confirmMealScanItems
      summary: Confirm foods and portions; returns calculated analysis.
      tags: [meals]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmItemsRequest' }
      responses:
        '200':
          description: Calculated analysis. Not a diary log.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MealAnalysisResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /v1/meal-logs:
    post:
      operationId: createMealLog
      summary: Log a confirmed meal.
      tags: [meals]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMealLogRequest' }
      responses:
        '201':
          description: Logged meal snapshot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MealLogResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      operationId: listMealLogs
      summary: Meal diary for a date.
      tags: [meals]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/DateQueryRequired'
      responses:
        '200':
          description: Diary for the date with coverage and totals.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MealDiaryResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/meal-logs/{id}:
    patch:
      operationId: patchMealLog
      summary: Edit a meal log.
      tags: [meals]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchMealLogRequest' }
      responses:
        '200':
          description: Changed log.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MealLogResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deleteMealLog
      summary: Delete a meal log.
      tags: [meals]
      x-noura-milestone: M4
      x-noura-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ExpectedRevisionQuery'
      responses:
        '200':
          description: Deleted; derived summaries invalidated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token. Verified via JWKS with issuer, audience, expiry and subject checks.
    revenueCatWebhookAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Shared authorization value configured in RevenueCat. Server secret; never in the app.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key (UUID recommended). Replays return the original result.
      schema: { type: string, minLength: 8, maxLength: 255 }
...
    DateQueryRequired:
      name: date
      in: query
      required: true
      schema: { type: string, format: date }
    ExpectedRevisionQuery:
      name: expected_revision
      in: query
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    Unauthenticated:
      description: Missing, invalid or expired access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: UNAUTHENTICATED, message: Authentication required., field_errors: [], retryable: false }
            meta: { request_id: 7b0f8c1e-3b8f-4a52-9d55-3d1d1f6a0c11 }
    NotFound:
      description: Resource does not exist or is not owned by the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: NOT_FOUND, message: Resource not found., field_errors: [], retryable: false }
            meta: { request_id: 7b0f8c1e-3b8f-4a52-9d55-3d1d1f6a0c11 }
    RevisionConflict:
      description: expected_revision is stale.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Invalid content.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed.
              field_errors: [{ field: body.weight_kg, code: maximum, message: must be <= 400 }]
              retryable: false
            meta: { request_id: 7b0f8c1e-3b8f-4a52-9d55-3d1d1f6a0c11 }
...
    QuotaExceeded:
      description: Rate or quota limit reached.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    Meta:
      type: object
      additionalProperties: false
      required: [request_id]
      properties:
        request_id: { type: string, minLength: 1, maxLength: 128 }
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - NOT_FOUND
        - REVISION_CONFLICT
        - QUOTA_EXCEEDED
        - PROVIDER_UNAVAILABLE
        - UNSUPPORTED_INPUT
        - CONSTRAINT_CONFLICT
        - INTERNAL_ERROR
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string, maxLength: 200 }
        code: { type: string, maxLength: 64 }
        message: { type: string, maxLength: 300 }
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, field_errors, retryable]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, maxLength: 500 }
        field_errors:
          type: array
          maxItems: 50
          items: { $ref: '#/components/schemas/FieldError' }
        retryable: { type: boolean }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    Deleted:
      type: object
      additionalProperties: false
      required: [id, deleted]
      properties:
        id: { type: string, format: uuid }
        deleted: { type: boolean }
    DeletedResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Deleted' }
        meta: { $ref: '#/components/schemas/Meta' }

    # ---------------------------------------------------------------- jobs
...
    SafeError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string, maxLength: 64 }
        message: { type: string, maxLength: 500 }
...
    NumberRange:
      type: object
      additionalProperties: false
      required: [min, max]
      properties:
        min: { type: number }
        max: { type: number }
...
    Nutrients:
      type: object
      additionalProperties: false
      description: Null means unknown. Unknown is never reported as zero.
      required: [energy_kcal, protein_g, carbohydrate_g, fat_g, fibre_g]
      properties:
        energy_kcal: { type: number, nullable: true, minimum: 0 }
        protein_g: { type: number, nullable: true, minimum: 0 }
        carbohydrate_g: { type: number, nullable: true, minimum: 0 }
        fat_g: { type: number, nullable: true, minimum: 0 }
        fibre_g: { type: number, nullable: true, minimum: 0 }
    Coverage:
      type: object
      additionalProperties: false
      required: [items_total, items_with_nutrition, complete]
      properties:
        items_total: { type: integer, minimum: 0 }
        items_with_nutrition: { type: integer, minimum: 0 }
        complete: { type: boolean }
    NutrientTotals:
      type: object
      additionalProperties: false
      required: [nutrients, coverage]
      properties:
        nutrients: { $ref: '#/components/schemas/Nutrients' }
        coverage: { $ref: '#/components/schemas/Coverage' }
    SourceRef:
      type: object
      additionalProperties: false
      required: [source_id, source_name, source_version]
      properties:
        source_id: { type: string, format: uuid }
        source_name: { type: string }
        source_version: { type: string }
    MealSlot:
      type: string
      enum: [breakfast, lunch, dinner, snack]
    Uncertainty:
      type: string
      enum: [low, medium, high]
    OilLevel:
      type: string
      enum: [none, light, medium, heavy, unknown]
    PreparationInput:
      type: object
      additionalProperties: false
      properties:
        method: { type: string, maxLength: 64 }
        oil_level: { $ref: '#/components/schemas/OilLevel' }
        source: { type: string, enum: [homemade, restaurant, packaged, unknown] }
    ServingInput:
      type: object
      additionalProperties: false
      required: [unit, quantity]
      properties:
        unit: { type: string, maxLength: 32 }
        quantity: { type: number, exclusiveMinimum: true, minimum: 0, maximum: 50 }
    ConfirmedItemInput:
      type: object
      additionalProperties: false
      description: Exactly one of grams or serving must be provided (validated server-side).
      required: [label]
      properties:
        temporary_id: { type: string, maxLength: 32 }
        food_id: { type: string, format: uuid, nullable: true }
        recipe_id: { type: string, format: uuid, nullable: true }
        label: { type: string, minLength: 1, maxLength: 200 }
        grams: { type: number, exclusiveMinimum: true, minimum: 0, maximum: 3000 }
        serving: { $ref: '#/components/schemas/ServingInput' }
        preparation: { $ref: '#/components/schemas/PreparationInput' }
    AnalyzedItem:
      type: object
      additionalProperties: false
      required: [item_id, label, food_id, recipe_id, source, grams, grams_range, nutrients, uncertainty]
      properties:
        item_id: { type: string, format: uuid }
        label: { type: string }
        food_id: { type: string, format: uuid, nullable: true }
        recipe_id: { type: string, format: uuid, nullable: true }
        source:
          type: object
          allOf: [{ $ref: '#/components/schemas/SourceRef' }]
          nullable: true
        grams: { type: number, nullable: true }
        grams_range:
          type: object
          allOf: [{ $ref: '#/components/schemas/NumberRange' }]
          nullable: true
        nutrients:
          type: object
          allOf: [{ $ref: '#/components/schemas/Nutrients' }]
          nullable: true
        uncertainty:
          type: string
          allOf: [{ $ref: '#/components/schemas/Uncertainty' }]
          nullable: true
    MealBalanceComponent:
      type: object
      additionalProperties: false
      required: [key, score, max_score, evidence]
      properties:
        key: { type: string, enum: [protein, fibre, vegetable_fruit, variety] }
        score: { type: number, nullable: true, minimum: 0, maximum: 25 }
        max_score: { type: integer, minimum: 25, maximum: 25, description: Each component is capped at 25 in policy v1. }
        evidence: { type: object, additionalProperties: true }
    MealBalance:
      type: object
      additionalProperties: false
      description: Explainable product heuristic, not a medical health score.
      required: [score, policy_version, components, missing_data_message]
      properties:
        score: { type: integer, nullable: true, minimum: 0, maximum: 100 }
        policy_version: { type: string }
        components:
          type: array
          maxItems: 4
          items: { $ref: '#/components/schemas/MealBalanceComponent' }
        missing_data_message: { type: string, nullable: true }

    # ---------------------------------------------------------------- home
...
    MediaPurpose:
      type: string
      enum: [meal, progress_photo]
    ImageMime:
      type: string
      enum: [image/jpeg, image/png, image/webp]
    UploadSlotRequest:
      type: object
      additionalProperties: false
      required: [purpose, mime, size_bytes]
      properties:
        purpose: { $ref: '#/components/schemas/MediaPurpose' }
        mime: { $ref: '#/components/schemas/ImageMime' }
        size_bytes: { type: integer, minimum: 1, maximum: 10485760 }
    UploadSlot:
      type: object
      additionalProperties: false
      required: [media_id, upload_url, expires_at]
      properties:
        media_id: { type: string, format: uuid }
        upload_url: { type: string, format: uri }
        expires_at: { type: string, format: date-time }
    UploadSlotResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/UploadSlot' }
        meta: { $ref: '#/components/schemas/Meta' }
    MediaAsset:
      type: object
      additionalProperties: false
      required: [id, purpose, status, verified_mime, byte_size, created_at]
      properties:
        id: { type: string, format: uuid }
        purpose: { type: string, enum: [meal, progress_photo, export] }
        status: { type: string, enum: [awaiting_upload, uploaded, verified, rejected, deleted] }
        verified_mime: { type: string, nullable: true }
        byte_size: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
    MediaAssetResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/MediaAsset' }
        meta: { $ref: '#/components/schemas/Meta' }
    MediaDownload:
      type: object
      additionalProperties: false
      required: [url, expires_at]
      properties:
        url: { type: string, format: uri }
        expires_at: { type: string, format: date-time }
    MediaDownloadResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/MediaDownload' }
        meta: { $ref: '#/components/schemas/Meta' }
    CreateMealScanRequest:
      type: object
      additionalProperties: false
      required: [media_id]
      properties:
        media_id: { type: string, format: uuid }
    MealScanAccepted:
      type: object
      additionalProperties: false
      required: [job_id, scan_id]
      properties:
        job_id: { type: string, format: uuid }
        scan_id: { type: string, format: uuid }
    MealScanAcceptedResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/MealScanAccepted' }
        meta: { $ref: '#/components/schemas/Meta' }
    RecognitionItem:
      type: object
      additionalProperties: false
      required: [temporary_id, label, alternative_labels, confidence_band, estimated_grams, preparation_questions, needs_confirmation, catalog_candidates]
      properties:
        temporary_id: { type: string, maxLength: 32 }
        label: { type: string, maxLength: 200 }
        alternative_labels: { type: array, maxItems: 5, items: { type: string, maxLength: 200 } }
        confidence_band: { type: string, enum: [low, medium, high] }
        estimated_grams:
          type: object
          allOf: [{ $ref: '#/components/schemas/NumberRange' }]
          nullable: true
        preparation_questions: { type: array, maxItems: 5, items: { type: string, maxLength: 200 } }
        needs_confirmation: { type: boolean }
        catalog_candidates:
          type: array
          maxItems: 5
          items: { type: string, format: uuid }
    Recognition:
      type: object
      additionalProperties: false
      required: [schema_version, image_is_food, quality, items, clarification]
      properties:
        schema_version: { type: string, enum: ['1'] }
        image_is_food: { type: boolean }
        quality: { type: string, enum: [usable, blurry, too_dark, ambiguous, unusable] }
        items:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/RecognitionItem' }
        clarification: { type: string, nullable: true, maxLength: 300 }
    MealScanStatus:
      type: string
      enum: [awaiting_upload, queued, recognizing, needs_confirmation, ready, failed, cancelled, expired]
    MealScan:
      type: object
      additionalProperties: false
      required: [id, status, recognition, revision, error, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/MealScanStatus' }
        recognition:
          type: object
          allOf: [{ $ref: '#/components/schemas/Recognition' }]
          nullable: true
        revision: { type: integer, minimum: 1 }
        error:
          type: object
          allOf: [{ $ref: '#/components/schemas/SafeError' }]
          nullable: true
        created_at: { type: string, format: date-time }
    MealScanResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/MealScan' }
        meta: { $ref: '#/components/schemas/Meta' }
    ConfirmItemsRequest:
      type: object
      additionalProperties: false
      required: [expected_revision, items]
      properties:
        expected_revision: { type: integer, minimum: 1 }
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: '#/components/schemas/ConfirmedItemInput' }
    MealAnalysis:
      type: object
      additionalProperties: false
      required: [scan_id, revision, items, totals, meal_balance]
      properties:
        scan_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 1 }
        items:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/AnalyzedItem' }
        totals: { $ref: '#/components/schemas/NutrientTotals' }
        meal_balance: { $ref: '#/components/schemas/MealBalance' }
    MealAnalysisResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/MealAnalysis' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    CreateMealLogRequest:
      type: object
      additionalProperties: false
      required: [client_id, consumed_at, timezone, slot, items]
      properties:
        client_id: { type: string, format: uuid }
        consumed_at: { type: string, format: date-time }
        timezone: { type: string, minLength: 1, maxLength: 64 }
        slot: { $ref: '#/components/schemas/MealSlot' }
        scan_id: { type: string, format: uuid, nullable: true }
        plan_meal_id: { type: string, format: uuid, nullable: true }
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: '#/components/schemas/ConfirmedItemInput' }
    PatchMealLogRequest:
      type: object
      additionalProperties: false
      required: [expected_revision]
      minProperties: 2
      properties:
        expected_revision: { type: integer, minimum: 1 }
        consumed_at: { type: string, format: date-time }
        slot: { $ref: '#/components/schemas/MealSlot' }
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: '#/components/schemas/ConfirmedItemInput' }
    MealLog:
      type: object
      additionalProperties: false
      required: [id, client_id, consumed_at, local_date, slot, scan_id, plan_meal_id, items, totals, meal_balance, revision]
      properties:
        id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        consumed_at: { type: string, format: date-time }
        local_date: { type: string, format: date }
        slot: { $ref: '#/components/schemas/MealSlot' }
        scan_id: { type: string, format: uuid, nullable: true }
        plan_meal_id: { type: string, format: uuid, nullable: true }
        items:
          type: array
          items: { $ref: '#/components/schemas/AnalyzedItem' }
        totals: { $ref: '#/components/schemas/NutrientTotals' }
        meal_balance:
          type: object
          allOf: [{ $ref: '#/components/schemas/MealBalance' }]
          nullable: true
        revision: { type: integer, minimum: 1 }
    MealLogResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/MealLog' }
        meta: { $ref: '#/components/schemas/Meta' }
    MealDiary:
      type: object
      additionalProperties: false
      required: [date, logs, totals]
      properties:
        date: { type: string, format: date }
        logs:
          type: array
          items: { $ref: '#/components/schemas/MealLog' }
        totals: { $ref: '#/components/schemas/NutrientTotals' }
    MealDiaryResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/MealDiary' }
        meta: { $ref: '#/components/schemas/Meta' }

    # ---------------------------------------------------------------- diet plans
...